requirements.txt
boto3==1.42.68
botocore==1.42.68
jmespath==1.1.0
joblib==1.5.3
numpy==2.4.2
paho-mqtt==2.1.0
pandas==3.0.1
python-dateutil==2.9.0.post0
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.3
urllib3==2.6.3
